Add Navigation Bar to Home
