+++ src/styles/globals.css
@@ -1,4 +1,7 @@
-html,
+@tailwind base;
+@tailwind components;
+@tailwind utilities;
+/* html,
 body {
   padding: 0;
   margin: 0;
@@ -13,4 +16,4 @@ a {
 
 * {
   box-sizing: border-box;
-}
+} */
